Move IdentifierType from BusinessId to Identifier
