+++ app/src/main/java/org/simple/clinic/AppDatabase.kt
@@ -21,6 +21,7 @@ import org.simple.clinic.patient.PatientPhoneNumberType
 import org.simple.clinic.patient.PatientSearchResult
 import org.simple.clinic.patient.PatientStatus
 import org.simple.clinic.patient.SyncStatus
+import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.patient.recent.RecentPatient
 import org.simple.clinic.protocol.Protocol
 import org.simple.clinic.protocol.ProtocolDrug
@@ -69,7 +70,7 @@ import org.simple.clinic.util.UuidRoomTypeConverter
     InstantRoomTypeConverter::class,
     LocalDateRoomTypeConverter::class,
     UuidRoomTypeConverter::class,
-    BusinessId.IdentifierType.RoomTypeConverter::class,
+    Identifier.IdentifierType.RoomTypeConverter::class,
     BusinessId.MetaVersion.RoomTypeConverter::class
 )
 abstract class AppDatabase : RoomDatabase() {
